// src/generated/model/feeCategoryRequest.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { FeeCategoryRequestFrequency } from './feeCategoryRequestFrequency.ts';

export interface FeeCategoryRequest {
  /**
     * @minLength 0
     * @maxLength 100
     */
  name: string;
  description?: string;
  frequency: FeeCategoryRequestFrequency;
  defaultAmount?: number;
}