// src/generated/model/updateFeeCategoryRequest.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { UpdateFeeCategoryRequestFrequency } from './updateFeeCategoryRequestFrequency.ts';
import type { UpdateFeeCategoryRequestStatus } from './updateFeeCategoryRequestStatus.ts';

export interface UpdateFeeCategoryRequest {
  /**
     * @minLength 0
     * @maxLength 100
     */
  name: string;
  description?: string;
  frequency: UpdateFeeCategoryRequestFrequency;
  defaultAmount?: number;
  status: UpdateFeeCategoryRequestStatus;
}