// src/generated/model/feeCategoryResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { FeeCategoryResponseFrequency } from './feeCategoryResponseFrequency.ts';
import type { FeeCategoryResponseStatus } from './feeCategoryResponseStatus.ts';

export interface FeeCategoryResponse {
  id?: string;
  name?: string;
  description?: string;
  frequency?: FeeCategoryResponseFrequency;
  defaultAmount?: number;
  status?: FeeCategoryResponseStatus;
  createdAt?: string;
}